pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "threat-plane"
version = "0.1.0"
description = "A Geometric Framework for Modeling and Visualizing Organizational Cyber Risk"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = [
    "cybersecurity",
    "threat-modeling",
    "graph-neural-networks",
    "topological-data-analysis",
    "visualization",
    "machine-learning"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]

dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "networkx>=2.6.0",
]

[project.optional-dependencies]
full = [
    "torch>=1.10.0",
    "torch-geometric>=2.0.0",
    "umap-learn>=0.5.0",
    "hdbscan>=0.8.0",
    "gudhi>=3.5.0",
    "plotly>=5.0.0",
    "kmapper>=2.0.0",
]
viz = [
    "plotly>=5.0.0",
    "matplotlib>=3.5.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/yourusername/threat-plane"
Documentation = "https://github.com/yourusername/threat-plane#readme"
Repository = "https://github.com/yourusername/threat-plane"
Issues = "https://github.com/yourusername/threat-plane/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
